entre_agents/requirements.txt
# VEntre AI Agent System - Python Dependencies

# Core AI/LLM
anthropic>=0.18.0          # Anthropic Claude API client
openai>=1.12.0             # OpenAI API client (optional)

# Database
# SQLite is built-in to Python, no package needed
psycopg2-binary>=2.9.9     # PostgreSQL adapter (optional, only for cloud deployment)

# Configuration
python-dotenv>=1.0.0       # Load environment variables from .env files

# Utilities
typing-extensions>=4.9.0   # Type hints support for older Python versions

# Development/Testing (optional)
pytest>=7.4.0             # Testing framework
pytest-cov>=4.1.0         # Test coverage
black>=23.12.0            # Code formatter
flake8>=7.0.0             # Linter
